move utills function to a separate dispathcer file
